+++ src/gMathFunc.cpp
@@ -64,7 +64,7 @@ long double gamma(long double v) {
 
 /* Some special handling of mlv*mdv */
 mlv operator*(const mlv& mv1, const long double d) {
-	return mlv(	(mv1.x * (d * LACCURACY)) / LACCURACY, 
-			(mv1.y * (d * LACCURACY)) / LACCURACY, 
-			(mv1.z * (d * LACCURACY)) / LACCURACY);
+	return mlv(	(long long)((mv1.x * (d * LACCURACY)) / LACCURACY), 
+			(long long)((mv1.y * (d * LACCURACY)) / LACCURACY), 
+			(long long)((mv1.z * (d * LACCURACY)) / LACCURACY));
 }
